+++ BIM32/thingspeak.hpp
@@ -367,7 +367,7 @@ String Thingspeak::_fieldPrepare(unsigned int fieldNum) {
         float co2 = wsensor.get_co2(wsensNum, config.wsensor_co2_corr(wsensNum));
         float vbat = wsensor.get_batteryVoltage(wsensNum);
         int batlvl = wsensor.get_batteryLevel(wsensNum);
-        int batprc = wsensor.get_batteryPercentage(wsensNum);
+        float batprc = wsensor.get_batteryPercentage(wsensNum);
         
         if(wsensType >= 0 and wsensType <= 4 and sensors.checkTemp(temp)) fields = field + String(temp);
         if(wsensType == 5 and sensors.checkHum(hum)) fields = field + String(hum);
