+++ BIM32/wirelessSensor.hpp
@@ -11,7 +11,7 @@ class WirelessSensor {
     bool checkCo2(float co2);
     bool checkBatVolt(float volt);
     bool checkBatLvl(int lvl);
-    bool checkBatPercent(int percent);
+    bool checkBatPercent(float percent);
     unsigned int get_updated(unsigned int num);
     float get_temperature(unsigned int num, unsigned int sensor, float corr);
     String get_sensorType(unsigned int num, unsigned int sensor);
@@ -30,7 +30,7 @@ class WirelessSensor {
     int get_batteryAdc(unsigned int num);
     float get_batteryVoltage(unsigned int num);
     int get_batteryLevel(unsigned int num);
-    int get_batteryPercentage(unsigned int num);
+    float get_batteryPercentage(unsigned int num);
 
   private:
     unsigned int _updated[WSENSORS] = {0, 0};        // Last update timestamp
@@ -52,7 +52,7 @@ class WirelessSensor {
     int _adc[WSENSORS] = {-1, -1};                   // Battery ADC raw data
     float _batteryVoltage[WSENSORS] = {-1.0, -1.0};  // Battery voltage
     int _batteryLevel[WSENSORS] = {-1, -1};          // Battery level
-    int _batteryPercentage[WSENSORS] = {-1, -1};     // Battery percentage
+    float _batteryPercentage[WSENSORS] = {-1, -1};     // Battery percentage
 };
 
 void WirelessSensor::receive() {
@@ -107,8 +107,8 @@ void WirelessSensor::receive() {
           else if(_batteryVoltage[number] < (umin + stp * 2)) _batteryLevel[number] = 2;
           else if(_batteryVoltage[number] < (umin + stp * 3)) _batteryLevel[number] = 3;
           else _batteryLevel[number] = 4;
-          _batteryPercentage[number] = (_batteryVoltage[number] - umin) * 100 / (umax - umin); 
-          if(_batteryPercentage[number] > 100) _batteryPercentage[number] = 100;
+          _batteryPercentage[number] = (_batteryVoltage[number] - umin) * 100.0 / (umax - umin); 
+          if(_batteryPercentage[number] > 100.0) _batteryPercentage[number] = 100.0;
         }
             
         Serial.printf("Sensor %d updated at: ", number);
@@ -192,8 +192,8 @@ bool WirelessSensor::checkBatLvl(int lvl) {
 /**
  * Check if battery percentage is within the normal range
  */
-bool WirelessSensor::checkBatPercent(int percent) {
-  return (percent >= 0 and percent <= 100);
+bool WirelessSensor::checkBatPercent(float percent) {
+  return (percent >= 0.0 and percent <= 100.0);
 }
         
 /*
@@ -291,7 +291,7 @@ int WirelessSensor::get_batteryLevel(unsigned int num) {
   return _batteryLevel[num];
 }
 
-int WirelessSensor::get_batteryPercentage(unsigned int num) {
-  if(num >= WSENSORS) return -1;
+float WirelessSensor::get_batteryPercentage(unsigned int num) {
+  if(num >= WSENSORS) return -1.0;
   return _batteryPercentage[num];
 }
